+++ cli/util/util.go
@@ -435,6 +435,12 @@ func IsURL(str string) bool {
 	return err == nil && u.Scheme != "" && u.Host != "" && u.Opaque == "" && u.User == nil
 }
 
+// IsAuthURL checks if str is a valid URL with authentication information.
+func IsAuthURL(str string) bool {
+	u, err := url.Parse(str)
+	return err == nil && u.Scheme != "" && u.Host != "" && u.User != nil
+}
+
 // removeScheme removes the scheme from the input URL.
 func RemoveScheme(inputURL string) (string, error) {
 	parsedURL, err := url.Parse(inputURL)